Flash/messages

+++ app/controllers/camp_admin/camplocations_controller.rb
@@ -5,13 +5,13 @@ module CampAdmin
   # comment
   class CamplocationsController < ApplicationController
     before_action :set_camp, only: %i[edit show update destroy]
-    helper_method :sort_column, :sort_direction
     before_action :authenticate_user!, except: %i[index show]
+    helper_method :sort_column, :sort_direction
 
     def index
       @camplocations = Camplocation.search(params[:p])
-      .order("#{sort_column} #{sort_direction}")
-      .page(params[:page])
+                                   .order("#{sort_column} #{sort_direction}")
+                                   .page(params[:page])
 
       respond_to do |format|
         format.html
@@ -21,32 +21,37 @@ module CampAdmin
 
     def show; end
 
-    def edit; end
-
     def new
       @camplocation = Camplocation.new
+      authorize([:camp_admin, @camplocation])
+    end
+
+    def edit
+      authorize([:camp_admin, @camplocation])
     end
 
     def create
       @camplocation = Camplocation.new(camp_params)
-      autherize @camplocation
-
+      authorize([:camp_admin, @camplocation])
       if @camplocation.save!
-        redirect_to camp_admin_camplocations_path
+        redirect_to camp_admin_camplocations_path, notice: "Your camplocation has been saved"
+
       else
-        render :new, status: :unprocessable_entity
+        render :new, alert: "Something went wrong"
       end
     end
 
     def update
+      authorize([:camp_admin, @camplocation])
       if @camplocation.update(camp_params)
-        redirect_to camp_admin_camplocations_path
+        redirect_to camp_admin_camplocations_path, notice: "Your profile has been updated."
       else
-        render 'edit'
+        render 'edit', alert: "Something went wrong"
       end
     end
 
     def destroy
+      authorize([:camp_admin, @camplocation])
       return unless @camplocation.destroy
 
       redirect_to camp_admin_camplocations_path
